requirements.txt
aiofiles==0.3.2
aiohttp==2.3.9
astroid==1.6.1
async-timeout==2.0.0
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
coverage==4.5.1
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.13
httptools==0.0.9
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
locustio==0.8.1
MarkupSafe==1.0
mccabe==0.6.1
motor==1.2.1
msgpack-python==0.5.5
multidict==4.1.0
pluggy==0.6.0
py==1.5.2
pydash==4.2.1
pylint==1.8.2
pymongo==3.6.0
pytest==3.4.0
pytest-cov==2.5.1
python-dotenv==0.6.1
pyzmq==17.0.0
requests==2.18.4
sanic==0.7.0
six==1.11.0
ujson==1.35
urllib3==1.22
uvloop==0.8.1
websockets==4.0.1
Werkzeug==0.14.1
wrapt==1.10.11
yarl==1.1.1
